wip: use nvim highlight api
